+++ lightnode.go
@@ -118,8 +118,19 @@ func New(options Options, ctx context.Context, logger logrus.FieldLogger, sqlDB 
 		}
 	}
 	verifier := resolver.NewVerifier(hostChains, verifierBindings)
+<<<<<<< HEAD
 	resolverI := resolver.New(logger, cacher, multiStore, db, serverOptions, compatStore, bindings, verifier)
 	server := jsonrpc.NewServer(serverOptions, resolverI, resolver.NewValidator(options.Network, verifierBindings, options.DistPubKey, compatStore, logger))
+=======
+	resolverI := resolver.New(options.Network, logger, cacher, multiStore, db, serverOptions, compatStore, bindings, verifier)
+	limiter := resolver.NewRateLimiter(resolver.RateLimiterConf{
+		GlobalMethodRate: options.LimiterGlobalRates,
+		IpMethodRate:     options.LimiterIPRates,
+		Ttl:              options.LimiterTTL,
+		MaxClients:       options.LimiterMaxClients,
+	})
+	server := jsonrpc.NewServer(serverOptions, resolverI, resolver.NewValidator(verifierBindings, options.DistPubKey, compatStore, &limiter, logger))
+>>>>>>> release/0.4.5
 	confirmer := confirmer.New(
 		confirmer.DefaultOptions().
 			WithLogger(logger).